feat: add CodeForces problem sets support and UI integration
- Add new CodeForces problem sets service to fetch and update problems
- Integrate CodeForces UI components including toolbar actions and views
- Add download icon for problem sets update action
- Modify database schema and models to support CodeForces problems
- Add tests for CodeForces API integration
- Update plugin configuration to include CodeForces settings

+++ src/main/scala/com/wenjunhuang/codeepiphany/codeforces/models/CodeForcesProblem.scala
@@ -3,21 +3,22 @@ package com.wenjunhuang.codeepiphany.codeforces.models
 import io.circe.derivation.ConfiguredDecoder
 
 case class CodeForcesProblem(
-  contestId: Option[Int],
+  contestId: Option[Long],
   problemsetName: Option[String],
   index: String,
   name: String,
   `type`: String,
-  points: Option[Double],
+  points: Option[Float],
   rating: Option[Int],
   tags: List[String]
 ) derives ConfiguredDecoder
 
-case class CodeForcesProblemStatistics(contestId: Option[Int], index: String, solvedCount: Int) derives ConfiguredDecoder
+case class CodeForcesProblemStatistics(contestId: Option[Long], index: String, solvedCount: Int)
+    derives ConfiguredDecoder
 
 case class CodeForcesProblemResult(
-                                    problems: List[CodeForcesProblem],
-                                    problemStatistics: List[CodeForcesProblemStatistics]
+  problems: List[CodeForcesProblem],
+  problemStatistics: List[CodeForcesProblemStatistics]
 ) derives ConfiguredDecoder
 
 case class CodeForcesProblemResponse(result: CodeForcesProblemResult) derives ConfiguredDecoder
